% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_simulate.R
\docType{data}
\name{acar_df_sim}
\alias{acar_df_sim}
\title{My Example Data}
\format{
A data frame with 100 rows and 7 variables:
\describe{
\item{Y}{Ordinal variable representing satisfaction level}
\item{X1}{Numeric variable generated from normal distribution}
\item{X2}{Numeric variable generated from uniform distribution}
\item{X3}{Numeric variable generated from beta distribution}
\item{X4}{Binary variable generated from binomial distribution}
\item{X5}{Factor variable generated from sample}
\item{X6}{Character variable generated from sample}
}
}
\source{
{Created to serve as an eaxample.}
}
\usage{
acar_df_sim
}
\description{
Description
This is an example dataset generated by the simulate_data function.
}
\examples{
data(acar_df)  # Lazy loading. Data becomes visibles ASAP
}
\keyword{datasets}
