% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison_test.R
\name{comparison_test}
\alias{comparison_test}
\title{Title}
\usage{
comparison_test(model_poly_autoregressive1, model_poly_autoregressive2)
}
\arguments{
\item{model_poly_autoregressive2}{}
}
\description{
Title
}
