% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison_test_dependent.R
\name{comparison_test_dependent}
\alias{comparison_test_dependent}
\title{Title}
\usage{
comparison_test_dependent(
  model_poly_autoregressive1,
  model_poly_autoregressive2
)
}
\arguments{
\item{model_poly_autoregressive2}{}
}
\description{
Title
}
